Refactor namespaces: split 'Internal' by functional role
Differentiate types by 'Implementation' vs 'Infrastructure' instead of
by access modifier.
