+++ XO.Console.Cli.Benchmarks/XO.Console.Cli.Benchmarks.csproj
@@ -9,8 +9,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\XO.Console.Cli\Internal\DefaultTypeResolver.cs" Link="Internal\DefaultTypeResolver.cs" />
-    <Compile Include="..\XO.Console.Cli\Internal\DefaultTypeResolverScope.cs" Link="Internal\DefaultTypeResolverScope.cs" />
+    <Compile Include="..\XO.Console.Cli\Implementation\DefaultTypeResolver.cs" Link="Implementation\DefaultTypeResolver.cs" />
+    <Compile Include="..\XO.Console.Cli\Implementation\DefaultTypeResolverScope.cs" Link="Implementation\DefaultTypeResolverScope.cs" />
   </ItemGroup>
 
   <ItemGroup>
@@ -21,10 +21,6 @@
   <ItemGroup>
     <ProjectReference Include="..\XO.Console.Cli.Extensions\XO.Console.Cli.Extensions.csproj" />
     <ProjectReference Include="..\XO.Console.Cli\XO.Console.Cli.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Internal\" />
   </ItemGroup>
 
 </Project>
