+++ XO.Console.Cli.Tests/CommandContextTest.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using XO.Console.Cli.Commands;
 using XO.Console.Cli.Features;
+using XO.Console.Cli.Implementation;
 using XO.Console.Cli.Model;
 using Xunit;
