+++ XO.Console.Cli.Tests/CommandOptionTest.cs
@@ -1,4 +1,5 @@
 using System;
+using XO.Console.Cli.Infrastructure;
 using XO.Console.Cli.Model;
 using Xunit;
