+++ XO.Console.Cli.Tests/CommandParametersInspectorTest.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using XO.Console.Cli.Implementation;
+using XO.Console.Cli.Infrastructure;
 using XO.Console.Cli.Model;
 using Xunit;
