+++ XO.Console.Cli/CommandAppBuilder.cs
@@ -1,6 +1,8 @@
 using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
+using XO.Console.Cli.Implementation;
+using XO.Console.Cli.Infrastructure;
 using XO.Console.Cli.Middleware;
 using XO.Console.Cli.Model;
