+++ XO.Console.Cli/Infrastructure/CommandParametersBinder.cs
@@ -1,7 +1,7 @@
 using System.Collections.Immutable;
 using XO.Console.Cli.Model;
 
-namespace XO.Console.Cli;
+namespace XO.Console.Cli.Infrastructure;
 
 internal sealed class CommandParametersBinder
 {
