+++ XO.Console.Cli/Infrastructure/CommandParametersInspector.cs
@@ -4,7 +4,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using XO.Console.Cli.Model;
 
-namespace XO.Console.Cli;
+namespace XO.Console.Cli.Infrastructure;
 
 internal sealed class CommandParametersInspector
 {
